+++ packages/e2e/cypress/e2e/api/spacePermissions.cy.ts
@@ -25,6 +25,7 @@ const dashboardBody = {
     name: 'private dashboard',
     description: '',
     tiles: [],
+    tabs: [],
 };
 
 const createPrivateChart = (