+++ packages/e2e/cypress/e2e/app/settings/profile.cy.ts
@@ -1,31 +1,27 @@
 import { SEED_ORG_1_ADMIN, SEED_ORG_1_ADMIN_EMAIL } from '@lightdash/common';
 
+const resetUserName = () => {
+    // Reset to default values
+
+    cy.request({
+        url: 'api/v1/user/me',
+        method: 'PATCH',
+        body: {
+            firstName: SEED_ORG_1_ADMIN.first_name,
+            lastName: SEED_ORG_1_ADMIN.last_name,
+            email: SEED_ORG_1_ADMIN_EMAIL.email,
+        },
+    });
+};
+
 describe('Settings - Profile', () => {
     beforeEach(() => {
         cy.login();
 
-        // Reset to default values
-        cy.request({
-            url: 'api/v1/user/me',
-            method: 'PATCH',
-            body: {
-                firstName: SEED_ORG_1_ADMIN.first_name,
-                lastName: SEED_ORG_1_ADMIN.last_name,
-                email: SEED_ORG_1_ADMIN_EMAIL.email,
-            },
-        });
+        resetUserName();
     });
     afterEach(() => {
-        // Reset to default values
-        cy.request({
-            url: 'api/v1/user/me',
-            method: 'PATCH',
-            body: {
-                firstName: SEED_ORG_1_ADMIN.first_name,
-                lastName: SEED_ORG_1_ADMIN.last_name,
-                email: SEED_ORG_1_ADMIN_EMAIL.email,
-            },
-        });
+        resetUserName();
     });
 
     it('should update user names', () => {
@@ -33,26 +29,27 @@ describe('Settings - Profile', () => {
         cy.findByTestId('user-avatar').click();
         cy.findByRole('menuitem', { name: 'User settings' }).click();
 
-        cy.findByPlaceholderText('First name')
-            .should('not.be.disabled')
-            .should('have.value', SEED_ORG_1_ADMIN.first_name);
-        cy.findByPlaceholderText('Last name')
-            .should('not.be.disabled')
-            .should('have.value', SEED_ORG_1_ADMIN.last_name);
         cy.findByPlaceholderText('Email')
             .should('not.be.disabled')
             .should('have.value', SEED_ORG_1_ADMIN_EMAIL.email);
 
+        cy.wait(500);
         cy.findByPlaceholderText('First name')
-            .focus()
+            .should('not.be.disabled')
+            .should('have.value', SEED_ORG_1_ADMIN.first_name)
+            .click()
             .clear()
-            .should('have.value', '');
-        cy.findByPlaceholderText('First name').type('Kevin');
+            .type('Kevin')
+            .blur();
+        cy.wait(500); // Add wait here to prevent flakiness when typing in different inputs.
         cy.findByPlaceholderText('Last name')
-            .focus()
+            .should('not.be.disabled')
+            .should('have.value', SEED_ORG_1_ADMIN.last_name)
+            .click()
             .clear()
-            .should('have.value', '');
-        cy.findByPlaceholderText('Last name').type('Space');
+            .type('Space')
+            .blur();
+        cy.wait(500);
 
         cy.findByRole('button', { name: 'Update' }).click();