+++ packages/frontend/src/components/DashboardTiles/AddTileButton.tsx
@@ -1,4 +1,8 @@
-import { DashboardTileTypes, type Dashboard } from '@lightdash/common';
+import {
+    DashboardTileTypes,
+    FeatureFlags,
+    type Dashboard,
+} from '@lightdash/common';
 import {
     Button,
     Group,
@@ -11,9 +15,11 @@ import {
     IconChartBar,
     IconInfoCircle,
     IconMarkdown,
+    IconNewSection,
     IconPlus,
     IconVideo,
 } from '@tabler/icons-react';
+import { useFeatureFlagEnabled } from 'posthog-js/react';
 import { useCallback, useState, type FC } from 'react';
 import { useHistory, useParams } from 'react-router-dom';
 import useDashboardStorage from '../../hooks/dashboard/useDashboardStorage';
@@ -24,13 +30,16 @@ import { TileAddModal } from './TileForms/TileAddModal';
 
 type Props = {
     onAddTiles: (tiles: Dashboard['tiles'][number][]) => void;
+    setAddingTab: (value: React.SetStateAction<boolean>) => void;
 } & Pick<ButtonProps, 'disabled'>;
 
-const AddTileButton: FC<Props> = ({ onAddTiles, disabled }) => {
+const AddTileButton: FC<Props> = ({ onAddTiles, setAddingTab, disabled }) => {
     const [addTileType, setAddTileType] = useState<DashboardTileTypes>();
     const [isAddChartTilesModalOpen, setIsAddChartTilesModalOpen] =
         useState<boolean>(false);
-
+    const isDashboardTabsEnabled = useFeatureFlagEnabled(
+        FeatureFlags.DashboardTabs,
+    );
     const dashboardTiles = useDashboardContext((c) => c.dashboardTiles);
     const dashboardFilters = useDashboardContext((c) => c.dashboardFilters);
     const haveTilesChanged = useDashboardContext((c) => c.haveTilesChanged);
@@ -117,6 +126,14 @@ const AddTileButton: FC<Props> = ({ onAddTiles, disabled }) => {
                     >
                         Loom video
                     </Menu.Item>
+                    {isDashboardTabsEnabled && (
+                        <Menu.Item
+                            onClick={() => setAddingTab(true)}
+                            icon={<MantineIcon icon={IconNewSection} />}
+                        >
+                            Add tab
+                        </Menu.Item>
+                    )}
                 </Menu.Dropdown>
             </Menu>