+++ packages/frontend/src/hooks/useQueryResults.tsx
@@ -9,7 +9,6 @@ import {
     type SortField,
 } from '@lightdash/common';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import dayjs from 'dayjs';
 import { useCallback, useMemo } from 'react';
 import { useParams } from 'react-router-dom';
 import { lightdashApi } from '../api';
@@ -19,6 +18,7 @@ import {
     convertDateFilters,
 } from '../utils/dateFilter';
 import useToaster from './toaster/useToaster';
+import useQueryError from './useQueryError';
 
 type QueryResultsProps = {
     projectUuid: string;
@@ -83,7 +83,7 @@ const getQueryResults = async ({
     const timezoneFixQuery = query && {
         ...query,
         filters: convertDateFilters(query.filters),
-        timezone: dayjs.tz.guess(),
+        timezone: query.timezone ?? undefined,
     };
 
     return lightdashApi<ApiQueryResults>({
@@ -103,7 +103,8 @@ export const useQueryResults = (props?: {
     dateZoomGranularity?: DateGranularity;
 }) => {
     const { projectUuid } = useParams<{ projectUuid: string }>();
-    const { showToastError } = useToaster();
+    const setErrorResponse = useQueryError();
+
     const fetchQuery =
         props?.isViewOnly === true ? getChartResults : getQueryResults;
     const mutation = useMutation<ApiQueryResults, ApiError, QueryResultsProps>(
@@ -111,10 +112,7 @@ export const useQueryResults = (props?: {
         {
             mutationKey: ['queryResults'],
             onError: (error) => {
-                showToastError({
-                    title: 'Error running query',
-                    subtitle: error.error.message,
-                });
+                setErrorResponse(error);
             },
         },
     );