+++ packages/frontend/src/hooks/useSavedQuery.ts
@@ -16,7 +16,6 @@ import {
     type UseMutationOptions,
     type UseQueryOptions,
 } from '@tanstack/react-query';
-import dayjs from 'dayjs';
 import { useHistory, useParams } from 'react-router-dom';
 import { lightdashApi } from '../api';
 import { convertDateFilters } from '../utils/dateFilter';
@@ -93,7 +92,7 @@ const addVersionSavedQuery = async ({
         metricQuery: {
             ...payload.metricQuery,
             filters: convertDateFilters(payload.metricQuery.filters),
-            timezone: dayjs.tz.guess(),
+            timezone: payload.metricQuery.timezone ?? undefined,
         },
     };
     return lightdashApi<SavedChart>({